+++ CHANGELOG.md
@@ -17,3 +17,12 @@
 ## 1.0.0
 
 - stable version
+
+## 1.1.0
+
+- add somes features as call custom fonction to execute when the request is unauthorized by adding directly the function in the instanciation
+- no need to import the error type file to get api error response
+
+## 1.1.0
+
+- update example